Resolved CID 351737 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument this->_accountMap.size() to format specifier %u was expected to have type unsigned int but has type unsigned long

+++ src/logonserver/Server/AccountMgr.cpp
@@ -210,7 +210,7 @@ void AccountMgr::reloadAccounts(bool silent)
     }
 
     if (!silent)
-        LogDefault("[AccountMgr] Found %u accounts.", _accountMap.size());
+        LogDefault("[AccountMgr] Found %u accounts.", static_cast<uint32_t>(_accountMap.size()));
 
     accountMgrMutex.Release();
 }
